+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/__init__.py
@@ -30,7 +30,11 @@ from ._models import (
     AccessPolicy,
     DelimitedTextDialect,
     DelimitedJsonDialect,
-    DataLakeFileQueryError
+    DataLakeFileQueryError,
+    AccessControlChangeResult,
+    AccessControlChangeCounters,
+    AccessControlChangeFailure,
+    AccessControlChanges,
 )
 from ._shared_access_signature import generate_account_sas, generate_file_system_sas, generate_directory_sas, \
     generate_file_sas
@@ -63,6 +67,10 @@ __all__ = [
     'PathPropertiesPaged',
     'LeaseProperties',
     'ContentSettings',
+    'AccessControlChangeResult',
+    'AccessControlChangeCounters',
+    'AccessControlChangeFailure',
+    'AccessControlChanges',
     'AccountSasPermissions',
     'FileSystemSasPermissions',
     'DirectorySasPermissions',